Update macro and function regex
Macros do not necessarily have to be followed by an `(` parenthesis, but
they *do* necessarily have to *not* be followed by any of the characters
inside the backticks: `|^/&<>~.=!*+-`. Because if they do, then they're
no longer normal macros, but _operators_.

From https://hexdocs.pm/elixir/operators.html

> To define an operator, you can use the usual def* constructs (def,
> defp, defmacro, and so on) but with a syntax similar to how the
> operator is used

So, what applied to `defmacro` applies equally to `def`, that's why we
added the bracket expression to the function regex.